chore(historical-tables): move preprocess-annotations to seaprate file
